new chnages relist and crom job

+++ app/Http/Controllers/Transporter/DashboardController.php
@@ -1220,9 +1220,11 @@ class DashboardController extends WebController
     public function saveSearchDlt(Request $request)
     {
         try {
+            if ($request->id){
             $data = SaveSearch::find($request->id);
             $data->delete();
             return redirect()->back()->with('saveSearchSuccess', 'Item deleted successfully');
+            }
         } catch (\Exception $ex) {
             return response(["success" => false, "message" => $ex->getMessage(), "data" => []]);
         }
